Fix DataManager Compiler Warnings

+++ src/DataManager/Lines/Line_Data.cpp
@@ -35,8 +35,8 @@ bool LineData::clearAtTime(TimeFrameIndex const time, int const line_id, bool no
     if (clear_at_time(time, line_id, _data)) {
         auto it = _entity_ids_by_time.find(time);
         if (it != _entity_ids_by_time.end()) {
-            if (line_id < static_cast<int>(it->second.size())) {
-                it->second.erase(it->second.begin() + line_id);
+            if (static_cast<size_t>(line_id) < it->second.size()) {
+                it->second.erase(it->second.begin() + static_cast<long int>(line_id));
             }
             if (it->second.empty()) {
                 _entity_ids_by_time.erase(it);
@@ -92,8 +92,8 @@ void LineData::addAtTime(TimeFrameIndex const time, Line2D const & line, bool no
 
 void LineData::addPointToLine(TimeFrameIndex const time, int const line_id, Point2D<float> point, bool notify) {
 
-    if (line_id < _data[time].size()) {
-        _data[time][line_id].push_back(point);
+    if (static_cast<size_t>(line_id) < _data[time].size()) {
+        _data[time][static_cast<size_t>(line_id)].push_back(point);
     } else {
         std::cerr << "LineData::addPointToLine: line_id out of range" << std::endl;
         _data[time].emplace_back();
@@ -107,15 +107,15 @@ void LineData::addPointToLine(TimeFrameIndex const time, int const line_id, Poin
 
 void LineData::addPointToLineInterpolate(TimeFrameIndex const time, int const line_id, Point2D<float> point, bool notify) {
 
-    if (line_id >= _data[time].size()) {
+    if (static_cast<size_t>(line_id) >= _data[time].size()) {
         std::cerr << "LineData::addPointToLineInterpolate: line_id out of range" << std::endl;
         _data[time].emplace_back();
     }
 
-    Line2D & line = _data[time][line_id];
+    Line2D & line = _data[time][static_cast<size_t>(line_id)];
     if (!line.empty()) {
         Point2D<float> const last_point = line.back();
-        double const distance = std::sqrt(std::pow(point.x - last_point.x, 2) + std::pow(point.y - last_point.y, 2));
+        float const distance = std::sqrt(std::pow(point.x - last_point.x, 2.0f) + std::pow(point.y - last_point.y, 2.0f));
         int const n = static_cast<int>(distance / 2.0f);
         for (int i = 1; i <= n; ++i) {
             float const t = static_cast<float>(i) / static_cast<float>(n + 1);
